backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.10.4
duckdb==1.1.0
gunicorn==22.0.0
